refactor: JS 비트 제한으로 ID String으로 변환

+++ src/main/kotlin/com/dh/baro/cart/application/CartFacade.kt
@@ -16,7 +16,7 @@ class CartFacade(
     fun addItem(userId: Long, request: AddItemRequest) {
         cartService.addItem(
             userId = userId,
-            productId = request.productId,
+            productId = request.productId.toLong(),
             quantity = request.quantity,
         )
     }
